LAMBDA/requirements.txt
azure-storage-queue 
azure-cosmos
azure-identity
azure-functions